avancement request & dashboard

+++ app/controllers/UserController.php
@@ -93,44 +93,51 @@ class UserController extends BaseController
 
 		if($validator->passes()){
 
-		$user = $location->building->user()->firstOrFail();
+			$user = $location->building->user()->firstOrFail();
 
-		if(isset($input['translate'])){ 
+			if(isset($input['translate'])){ 
 
-		$input['text'] = Helpers::translate($input['text'], Config::get('var.lang')[$user->language_id], Config::get('var.lang')[$user->language_id]);
+				$input['text'] = Helpers::translate($input['text'], Config::get('var.lang')[$user->language_id], Config::get('var.lang')[$user->language_id]);
+			}
+
+			$input['subject'] = Helpers::translate('Demande de réservation de votre bien n°:'.$location->id, 'fr', Config::get('var.lang')[$user->language_id]);
+
+			Mailgun::send('emails.reservation_location', array('input'=>$input,'user'=>$user,'location'=>$location,'sender'=>Auth::user()), function($message) use($input, $user)
+			{
+
+				$message
+				->from(Auth::user()->email, Auth::user()->first_name.' '.Auth::user()->name)
+				->to($user->email, $user->first_name.' '.$user->name)
+				->subject($input['subject']);
+
+			});
+
+			if(!$location->user()->whereUserId(Auth::user()->id)->whereStatus(1)->get()->count()){
+
+				if($location->user()->whereUserId(Auth::user()->id)->whereRequest(1)->get()->count()){
+
+					$location->user()->whereUserId(Auth::user()->id)->whereRequest(1)->detach();
+				}
+
+				$location->user()->attach(Auth::user(), array(
+					'begin'=>Helpers::dateNaForm($input['start_date']),
+					'end'=>$location->end_date,
+					'request'=>true,
+					'seat'=>$input['seat'],
+					'nb_locations'=>$input['nb_locations'],
+					'text'=>$input['text'],
+					));
+			}
+
+			return Redirect::back()
+			->withSuccess(trans('validation.custom.request_reservation_succes'));
+
+		}else{
+
+			return Redirect::back()
+			->withInput()
+			->withError($validator);
+		}
 	}
-
-	$input['subject'] = Helpers::translate('Demande de réservation de votre bien n°:'.$location->id, 'fr', Config::get('var.lang')[$user->language_id]);
-
-	Mailgun::send('emails.reservation_location', array('input'=>$input,'user'=>$user,'location'=>$location,'sender'=>Auth::user()), function($message) use($input, $user)
-	{
-
-	$message
-	->from(Auth::user()->email, Auth::user()->first_name.' '.Auth::user()->name)
-	->to($user->email, $user->first_name.' '.$user->name)
-	->subject($input['subject']);
-
-});
-
-if(!$location->user()->whereStatus(1)->get()->count()){
-
-if($location->user()->whereRequest(1)->get()->count()){
-
-$location->user()->whereRequest(1)->detach();
-}
-
-$location->user()->attach(Auth::user(), array('begin'=>Helpers::dateNaForm($input['start_date']),'end'=>$location->end_date,'request'=>true));
-}
-
-return Redirect::back()
-->withSuccess(trans('validation.custom.request_reservation_succes'));
-
-}else{
-
-return Redirect::back()
-->withInput()
-->withError($validator);
-}
-}
 
 }
